Fix : chat refactor
