+++ src/page/chat.tsx
@@ -1,4 +1,4 @@
-import { useState, useRef, useEffect } from "react";
+import { useState, useEffect } from "react";
 import chatAPI from "@/api/chat/chatAPI";
 import "@/components/styles/game-dashboard-animations.css";
 import { usePrivy } from "@privy-io/react-auth";
@@ -10,51 +10,9 @@ import ChatInput from "@/components/Chat/ChatInput";
 import signTransaction from "@/components/wallet/SignWallet";
 import Loading from "@/components/styles/spiner/wormhole/Loading";
 
-interface Chatting {
-  externalId?: string | null;
-  conversationExternalId?: string;
-  sender?: string | null;
-  content: string;
-  messageType: string;
-  data?: any | null;
-}
-
-interface GameRelation {
-  key: string;
-  content: string;
-}
-
-interface GameData {
-  gameTitle: string;
-  gameContent: string;
-  extras: string;
-  gameStartAt: string;
-  gameEndAt: string;
-  gameExpriedAt: string;
-  fixtureId: number;
-  gameRelations: GameRelation[];
-  quantity: string;
-  key: string;
-  asset: string;
-  choiceType: string;
-}
-
-interface ChatMessage {
-  externalId: string | null;
-  content: string;
-  messageType: "CREATE_TR";
-  data: GameData;
-}
-
-interface ConversationResponse {
-  status: string;
-  data: {
-    conversation: {
-      externalId: string;
-      messages: Chatting[];
-    };
-  };
-}
+import { Chatting, MessageData } from "@/types/chat";
+import { useChat } from "@/hooks/useChat";
+import { useWebSocket } from "@/hooks/useWebSocket";
 
 function ChattingComponents() {
   const { user } = usePrivy();
@@ -69,65 +27,21 @@ function ChattingComponents() {
   const homeInputText = location.state?.message || "";
   const wallet = JSON.parse(localStorage.getItem("user_wallet_info") || "{}");
   console.log("wallet", wallet);
-  const [messages, setMessages] = useState<Chatting[]>([]);
   const [inputText, setInputText] = useState<string>("");
-  const [loading, setLoading] = useState<boolean>(false);
-  const [conversationExternalId, setConversationExternalId] = useState<
-    string | null
-  >(null);
 
   const [bridgeLoading, setBridgeLoading] = useState<boolean>(false);
   const [bridge, setBridge] = useState<"yes" | "no">("no");
-  const isHomeMessageProcessed = useRef(false);
-  console.log("messages", messages);
-  const getWelcomeMessage = (username: string): Chatting => ({
-    externalId: null,
-    content: `
-#### Hello, **${username}**! 
-There are three options you can choose from: 
-`,
-    messageType: "Market_Info",
-    sender: "SYSTEM",
-    data: {
-      selections: [
-        {
-          name: "Create Prediction",
-          type: "option",
-          description: "Start your own prediction market.",
-        },
-        {
-          name: "Sports Search",
-          type: "option",
-          description: "Join an existing prediction market.",
-        },
-        {
-          name: "Chat",
-          type: "option",
-          description: "Ask PrediX anything you want to know!",
-        },
-      ],
-    },
-  });
 
-  const fetchConversationMessages = async (id: string) => {
-    try {
-      setLoading(true);
-      const response: ConversationResponse = await chatAPI.getChatMessages(id);
-      if (response.status === "SUCCESS") {
-        setMessages([
-          getWelcomeMessage(username),
-          ...response.data.conversation.messages,
-        ]);
-        setConversationExternalId(response.data.conversation.externalId);
-      }
-    } catch (error) {
-      console.error("Error fetching conversation messages:", error);
-      setMessages([getWelcomeMessage(username)]);
-    } finally {
-      setLoading(false);
-    }
-  };
-
+  const {
+    messages,
+    loading,
+    conversationExternalId,
+    setMessages,
+    setLoading,
+    sendChatMessage,
+    fetchConversationMessages,
+    getWelcomeMessage
+  } = useChat(username);
 
   useEffect(() => {
     if (externalId) {
@@ -135,32 +49,11 @@ There are three options you can choose from:
     } else {
       setMessages([getWelcomeMessage(username)]);
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [externalId]);
-  useEffect(() => {
-    chatAPI.connectSocket();
-
-    const interval = setInterval(() => {
-      if (chatAPI.connectSocket()) {
-        clearInterval(interval);
-
-        if (!isHomeMessageProcessed.current && homeInputText.trim() !== "") {
-          isHomeMessageProcessed.current = true;
-          const newMessage: Chatting = {
-            externalId: null,
-            content: homeInputText,
-            messageType: "TEXT",
-            sender: null,
-          };
-          sendChatMessage(newMessage);
-          navigate(location.pathname, { replace: true, state: {} });
-        }
-      }
-    }, 100);
-
-    return () => clearInterval(interval);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  useWebSocket(homeInputText, (message: Chatting) => {
+    sendChatMessage(message);
+    navigate(location.pathname, { replace: true, state: {} });
+  });
 
   useEffect(() => {
     const lastMarketMessage = messages.find(
@@ -173,34 +66,8 @@ There are three options you can choose from:
     if (lastMarketMessage && bridge.toLowerCase() === "yes") {
       BridgeCreateMessage();
     }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [messages]);
 
-  const sendChatMessage = async (message: Chatting) => {
-    setMessages((prevMessages) => [...prevMessages, message]);
-    setLoading(true);
-    try {
-      chatAPI.sendSoketMessage(message);
-    } catch (error) {
-      console.error("WebSocket 전송 실패:", error);
-    } finally {
-      chatAPI.addSocketListener((msg: Chatting) => {
-        const filteredContent = msg.content.replace(
-          /\n\s*-\s*\*\*Fixture ID:\*\*\s*\d+/g,
-          ""
-        );
-        setMessages((prevMessages) => [
-          ...prevMessages,
-          { ...msg, content: filteredContent.trim() },
-        ]);
-        setLoading(false);
-        if (!conversationExternalId && msg.conversationExternalId) {
-          setConversationExternalId(msg.conversationExternalId);
-        }
-      });
-    }
-  };
 
   const sendMessage = () => {
     if (inputText.trim() === "") return;
@@ -304,7 +171,7 @@ Great, I can fetch information related to sports. Currently, I only support foot
       if (!lastMarketMessage)
         throw new Error("No MARKET_FINALIZED message found");
 
-      const userGameSelectionData: ChatMessage = {
+      const userGameSelectionData: MessageData = {
         externalId: lastMarketMessage.conversationExternalId ?? null,
         content: lastMarketMessage.data?.event.league.name,
         messageType: "CREATE_TR",
